+++ api/moa.py
@@ -1,7 +1,5 @@
 import os
 from datetime import date, datetime
-import pandas as pd
-from typing import List
 import traceback
 from fastapi import FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
@@ -10,8 +8,7 @@ from schemas.IO import APIResponse , ChatRequest
 
 from langchain_openai import OpenAIEmbeddings
 from langchain_openai import ChatOpenAI
-from langchain_community.llms import OpenAI
-from api.util.util import *
+from api.retriever import MetadataFirstRetriever
 
 # LangChain imports
 from langchain_chroma import Chroma
@@ -19,18 +16,14 @@ from langchain.chains import RetrievalQA
 from langchain.prompts import (
     ChatPromptTemplate,
     SystemMessagePromptTemplate,
-    HumanMessagePromptTemplate,
-    PromptTemplate
+    HumanMessagePromptTemplate
 )
-from langchain.chains.summarize import load_summarize_chain
 
-from langchain.schema import BaseRetriever, Document
-from starlette.concurrency import run_in_threadpool
+from langchain.schema import Document
 
-# --- 환경 설정 ---
+
+# 
 OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
-
-
 if not OPENAI_API_KEY:
     raise RuntimeError("Environment variable OPENAI_API_KEY is required")
 
@@ -45,8 +38,7 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-# --- LangChain 설정 ---
-# 1) 프롬프트 템플릿
+# 프롬프트 템플릿
 SYSTEM_PROMPT = \
 """<ROLE>
 You are FestivalBot, a smart assistant for recommending local festivals and events in Korea.
@@ -89,90 +81,15 @@ HUMAN_PROMPT = """
 """
 
 
-# 한국어 요약용 프롬프트
-custom_summarize_prompt = PromptTemplate(
-    input_variables=["text"], 
-    template=
-"""
-다음 내용을 한국어로 간결하고 명확하게 요약해 주세요:
-
-{text}
-"""
-)
-
+# 시스템프롬프트, 유저 입력 프롬프트 인스턴스 생성
 chat_prompt = ChatPromptTemplate.from_messages([
     SystemMessagePromptTemplate.from_template(SYSTEM_PROMPT),
     HumanMessagePromptTemplate.from_template(HUMAN_PROMPT)
 ])
 
-# Retriever 커스텀
-# --- Custom Retriever: Metadata-first, then Embedding fallback ---
-class MetadataFirstRetriever(BaseRetriever):
-    vectordb: Chroma
-    k_meta:   int = 10
-    k_embed:  int = k_meta * 2
-
-    def __init__(self, vectordb, k_meta=3, k_embed=5):
-        # Pass fields to BaseRetriever for Pydantic validation
-        super().__init__(vectordb=vectordb, k_meta=k_meta, k_embed=k_embed)
-        self.vectordb = vectordb
-        self.k_meta   = k_meta
-        self.k_embed  = k_embed
-
-    async def aget_relevant_documents(self, query: str) -> List[Document]:
-        # Async interface calls sync method in executor
-        return await run_in_threadpool(self.get_relevant_documents, query)
-
-    def get_relevant_documents(self, query: str) -> List[Document]:
-        parsed      = parse_user_input(query)
-        meta_filter = build_meta_filter(parsed)
-        print(f"[Retrieval] meta_filter={meta_filter}, k_meta={self.k_meta}")
-        docs: List[Document] = []
-        
-        # 1) Metadata-first filtering if meta_filter exists
-        if meta_filter:
-            # Apply meta_filter to search_kwargs to filter documents by metadata
-            search_kwargs = {
-                "filter": meta_filter,  # Apply the metadata filter
-                "k": self.k_meta  # Number of results to return
-            }
-            retr_meta = self.vectordb.as_retriever(search_kwargs=search_kwargs)
-            try:
-                docs = retr_meta.get_relevant_documents(query)
-                print(f"[Retrieval] metadata match count={len(docs)}")
-            except Exception as e:
-                print(f"[Error] Metadata filtering failed: {e}")
-                # Proceed to fallback if metadata filtering fails
-        else:
-            print("[Retrieval] no metadata filter provided, skipping metadata filtering")
-
-        # 2) Fallback to embedding similarity if needed
-        if len(docs) < self.k_meta:
-            retr_embed = self.vectordb.as_retriever(
-                search_kwargs={"k": self.k_embed}
-            )
-            try:
-                more = retr_embed.get_relevant_documents(query)
-                print(f"[Retrieval] Embedding fallback retrieved={len(more)}")
-            except Exception as e:
-                print(f"[Error] Embedding fallback failed: {e}")
-                more = []
-            
-            # Merge without duplicates
-            seen = {d.metadata.get("event_id") for d in docs}
-            for d in more:
-                if d.metadata.get("event_id") not in seen:
-                    docs.append(d)
-                    if len(docs) >= self.k_meta:
-                        break
-                
-        print(f"[Retrieval] final docs count={len(docs)}")
-        return docs
-
 
 # catch vector
 embedding = OpenAIEmbeddings()
-
 vectordb = Chroma(
     persist_directory="/app/chroma_data",
     embedding_function=embedding,
@@ -189,8 +106,6 @@ def make_qa_chain(retriever):
         retriever=retriever,
         return_source_documents=True
     )
-
-# description 요약용 질의 체인
 
 # end point
 @app.post("/api/events", response_model=APIResponse)
